requirements.txt
docutils==0.12
flake8==2.5.4
mccabe==0.4.0
nose-parameterized==0.5.0
pep8==1.7.0
Pillow==3.1.1
py==1.4.31
pyflakes==1.0.0
Pygments==2.1.3
pytest==2.9.1
